LAST COMMIT I HOPE

+++ src/main/java/jeux/fousfous/heuristiques/DiffPions.java
@@ -4,10 +4,10 @@ import structure.*;
 
 public class DiffPions implements Heuristique{
 	@Override
-	public float estimate(PlateauFouFou plateau, State joueur){
-		float res = 0.f;
+	public int estimate(PlateauFouFou plateau, State joueur){
+		int res = 0;
 		for(int i = 0; i < PlateauFouFou.pSize * PlateauFouFou.pSize; i++){
-			res += plateau.getStateArray()[i] == joueur ? 0.f : -1.f;
+			res += plateau.getStateArray()[i] == joueur ? 0 : -1;
 		}
 		return res;
 	}
